pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "track-manager"
version = "0.1.0"
description = "Universal music downloader with smart duplicate detection and metadata management"
readme = "README.md"
requires-python = ">=3.8,<3.14"
license = {text = "MIT"}
authors = [
    {name = "Track Manager Contributors"}
]
keywords = ["music", "downloader", "spotify", "youtube", "soundcloud", "metadata", "duplicate-detection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    "yt-dlp>=2023.10.0",
    "spotdl>=4.2.0",
    "requests>=2.31.0",
    "mutagen>=1.47.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "bandit>=1.7.0",
]

[project.scripts]
track-manager = "track_manager.cli:main"
tm = "track_manager.cli:main"

[project.urls]
Homepage = "https://github.com/AmalganOpen/track-manager"
Documentation = "https://github.com/AmalganOpen/track-manager#readme"
Repository = "https://github.com/AmalganOpen/track-manager"
Issues = "https://github.com/AmalganOpen/track-manager/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["track_manager*"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--verbose --cov=track_manager --cov-report=html --cov-report=term-missing"
